refactor app into a routing component and move everything into FormPage
